+++ src/components/editors/GroupsEditor/components/GroupEditor/GroupEditor.tsx
@@ -448,7 +448,6 @@ export const GroupEditor: React.FC<Props> = ({ value, name, data, onChange, dash
                           dropdowns={dropdowns}
                           dashboards={dashboards}
                           isGrid={value.gridLayout}
-                          isHighlightTimePicker={value.highlightCurrentTimepicker}
                         />
                       </Collapse>
                     </div>