+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the **Business Links** panel are documented in this file.
 
 - Use raw values for time range in URL ([Issue #59](https://github.com/VolkovLabs/business-links/issues/59)).
 - Updated ESLint configuration ([#60](https://github.com/VolkovLabs/business-text/issues/60)).
+- Update sticky positioning for dynamic layouts ([#62](https://github.com/VolkovLabs/business-text/issues/62)).
 
 ## [2.1.0] - 2025-07-06