+++ CHANGELOG.md
@@ -7,9 +7,10 @@ All notable changes to the **Business Links** panel are documented in this file.
 ### Changed
 
 - Use raw values for time range in URL ([Issue #59](https://github.com/VolkovLabs/business-links/issues/59)).
-- Updated ESLint configuration ([#60](https://github.com/VolkovLabs/business-text/issues/60)).
-- Update sticky positioning for dynamic layouts ([#62](https://github.com/VolkovLabs/business-text/issues/62)).
-- Update Grid options for row height and width ([#63](https://github.com/VolkovLabs/business-text/issues/63)).
+- Updated ESLint configuration ([#60](https://github.com/VolkovLabs/business-links/issues/60)).
+- Updated sticky positioning for dynamic layouts ([#62](https://github.com/VolkovLabs/business-links/issues/62)).
+- Updated Grid options for row height and width ([#63](https://github.com/VolkovLabs/business-links/issues/63)).
+- Updated manual and relative time picker component to support additional time ranges ([#64](https://github.com/VolkovLabs/business-links/issues/64)).
 
 ## [2.1.0] - 2025-07-06