+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to the **Business Links** panel are documented in this file.
 
 ## [2.2.0] - Unreleased
 
+### Added
+
+- Added Grid options for row height and width ([#63](https://github.com/VolkovLabs/business-links/issues/63)).
+- Add MCP Servers functionality ([#56](https://github.com/VolkovLabs/business-links/issues/56)).
+
 ### Changed
 
 - Use raw values for time range in URL ([Issue #59](https://github.com/VolkovLabs/business-links/issues/59)).
 - Updated ESLint configuration ([#60](https://github.com/VolkovLabs/business-links/issues/60)).
 - Updated sticky positioning for dynamic layouts ([#62](https://github.com/VolkovLabs/business-links/issues/62)).
-- Updated Grid options for row height and width ([#63](https://github.com/VolkovLabs/business-links/issues/63)).
 - Updated manual and relative time picker component to support additional time ranges ([#64](https://github.com/VolkovLabs/business-links/issues/64)).
 
 ## [2.1.0] - 2025-07-06